pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-socketio>=5.5.1",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pyqt5>=5.15.11",
    "pyserial>=3.5",
    "qrcode>=8.2",
]
